requirements.txt
playwight
Twisted==22.10.0
scrapy-redis>=0.6.8
redis>=5.0
scrapy>=2.0
playwright-stealth>=1.0.5
beautifulsoup4>=4.12.2
python-dotenv


